Adding security

+++ src/main/java/com/exadelinternship/carpool/services/CarService.java
@@ -3,12 +3,13 @@ package com.exadelinternship.carpool.services;
 import com.exadelinternship.carpool.adapters.CarAdapter;
 import com.exadelinternship.carpool.dto.CarDTO;
 import com.exadelinternship.carpool.entity.Car;
-import com.exadelinternship.carpool.entity.User;
+import com.exadelinternship.carpool.entity.impl.UserDetailsImpl;
 import com.exadelinternship.carpool.repository.CarRepository;
-import com.exadelinternship.carpool.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -18,24 +19,29 @@ public class CarService {
     private CarRepository carRepository;
 
     @Autowired
-    private UserRepository userRepository;
-
-    @Autowired
     private CarAdapter carAdapter;
 
     public List<CarDTO> getAllCars(){
         List<CarDTO> result=new ArrayList<>();
-        carRepository.findAll().forEach(car -> result.add(carAdapter.carToCarDto(car)));
+        long userId=((UserDetailsImpl)SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getId();
+        carRepository.findAllByUser_Id(userId).forEach(car -> result.add(carAdapter.carToCarDto(car)));
         return result;
     }
 
+    @Transactional
     public void saveCar(CarDTO carDTO){
-        Car car=carAdapter.carDtoToCar(carDTO);
-        System.out.println("Info "+carDTO.getCarInformation());
+        long userId=((UserDetailsImpl)SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getId();
+        Car car=carAdapter.carDtoToCar(carDTO,userId);
         carRepository.save(car);
     }
 
-    public void deleteById(long id){
-        carRepository.deleteById(id);
+    @Transactional
+    public void deleteCarById(long id){
+        long userId=((UserDetailsImpl)SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getId();
+        Car car=carRepository.getOne(id);
+        if(car.getUser().getId()==userId) {
+            carRepository.deleteById(id);
+        }
+
     }
 }
